Extracted enroll to another controller
