+++ app/views/users/enrollment/users.json.jbuilder
