+++ config/routes.rb
@@ -124,7 +124,7 @@ Rails.application.routes.draw do
           constraints: { id: /.*/ }, via: [:post, :delete]
     match 'courses/*id/tag' => 'courses#tag',
           constraints: { id: /.*/ }, via: [:post, :delete]
-    match 'courses/*id/user' => 'users#enroll',
+    match 'courses/*id/user' => 'users/enrollment#index',
           constraints: { id: /.*/ }, via: [:post, :delete]
 
     # show-type actions: first all the specific json endpoints,
